chore: embed dashboard sdk code reorganization
