+++ packages/frontend/src/ee/features/embed/EmbedDashboard/styles/print.css
