+++ packages/frontend/src/ee/features/embed/EmbedDashboard/types.ts
@@ -1,0 +1,8 @@
+import { type FilterOperator } from '@lightdash/common';
+
+export type SdkFilter = {
+    model: string;
+    field: string;
+    operator: `${FilterOperator}`;
+    value: unknown | unknown[];
+};
