+++ packages/frontend/src/ee/providers/Embed/context.ts
@@ -3,6 +3,8 @@ import { type EmbedContext } from './types';
 
 const EmbedProviderContext = createContext<EmbedContext>({
     embedToken: undefined,
+    filters: undefined,
+    projectUuid: undefined,
 });
 
 export default EmbedProviderContext;
