+++ packages/frontend/src/ee/providers/Embed/types.ts
@@ -1,3 +1,7 @@
+import { type SdkFilter } from '../../features/embed/EmbedDashboard/types';
+
 export interface EmbedContext {
     embedToken?: string;
+    filters?: SdkFilter[];
+    projectUuid?: string;
 }
