+++ packages/sdk/src/index.ts
@@ -1,3 +1,1 @@
-import Lightdash from './Lightdash';
-export { FilterOperator } from '@lightdash/frontend';
-export default Lightdash;
+export { default } from './Lightdash';
